Move hashmap/hashset to toplevel
